refactor packagesapi and mtexamsapi
